feat: add user registration and password recovery functionality

+++ app/Http/Controllers/TaskController.php
@@ -12,8 +12,10 @@ class TaskController extends Controller
      *
      * @return JsonResponse
      */
-    public function all(): JsonResponse {
-        return response()->json(Task::with('completions')->orderByDesc('id')->get());
+    public function all(): JsonResponse
+    {
+        $tasks = Task::with('completions')->where('user_id', auth()->id())->orderByDesc('id')->get();
+        return response()->json($tasks);
     }
 
     /**
@@ -22,7 +24,8 @@ class TaskController extends Controller
      * @param Task $task - Task wildcard.
      * @return JsonResponse
      */
-    public function findById(Task $task): JsonResponse {
+    public function findById(Task $task): JsonResponse
+    {
         if (!$task) {
             return response()->setStatusCode(404)->json('Not found');
         }
@@ -36,7 +39,12 @@ class TaskController extends Controller
      * @param Task $task - Task wildcard.
      * @return JsonResponse
      */
-    public function delete(Task $task): JsonResponse {
+    public function delete(Task $task): JsonResponse
+    {
+        if (auth()->id() != $task->user_id) {
+            return response()->setStatusCode(401)->json('Unauthorized');
+        }
+
         try {
             return response()->json($task->delete());
         } catch (\Exception $e) {
@@ -50,7 +58,12 @@ class TaskController extends Controller
      * @param Task $task - Task wildcard.
      * @return JsonResponse
      */
-    public function update(Task $task): JsonResponse {
+    public function update(Task $task): JsonResponse
+    {
+        if (auth()->id() != $task->user_id) {
+            return response()->setStatusCode(401)->json('Unauthorized');
+        }
+
         if (!$task) {
             return response()->setStatusCode(404)->json('Not found');
         }
@@ -67,10 +80,12 @@ class TaskController extends Controller
      *
      * @return JsonResponse
      */
-    public function create(): JsonResponse {
+    public function create(): JsonResponse
+    {
         $task = new Task();
         $task->setAttribute('title', request()->get('title'));
         $task->setAttribute('frequency', request()->get('frequency'));
+        $task->setAttribute('user_id', auth()->id());
 
         try {
             $task->save();
